Files submitted for assignment.
These are the original files I submitted for a JavaScript assignment. I was asked to create a Rock, Paper, Scissors, Dynamite game where a user competes with the browser/computer. The user selects a move, then the browser randomly selects a move, and then the two moves are compared to determine a winner.